Address comments for PR#3
- Add more documentation to classes.
- Reorganize AndroidManifestUpdater to be an interface with 2018+
  and 2017 implementations.
